test: moved unit tests to /UnitTests/
