+++ test/UnitTests/operations/MailboxGetList.test.ts
@@ -1,8 +1,8 @@
 import { IExecuteFunctions } from 'n8n-workflow';
 import { ImapFlow } from 'imapflow';
-import { getMailboxListOperation } from '../../nodes/Imap/operations/mailbox/functions/MailboxGetList';
-import { createMockLogger } from '../TestUtils/N8nMocks';
-import { Imap } from '../../nodes/Imap/Imap.node';
+import { getMailboxListOperation } from '../../../nodes/Imap/operations/mailbox/functions/MailboxGetList';
+import { createMockLogger } from '../../TestUtils/N8nMocks';
+import { Imap } from '../../../nodes/Imap/Imap.node';
 
 describe('MailboxGetList', () => {
   let mockContext: jest.Mocked<IExecuteFunctions>;
